Move and add test for ChangesetEvents.Dedupe
